requirements.txt
arrow==0.15.2
aws-lambda-builders==0.4.0
aws-sam-cli==0.21.0
aws-sam-translator==1.11.0
awscli==1.16.238
binaryornot==0.4.4
boto3==1.9.228
botocore==1.12.228
certifi==2019.9.11
chardet==3.0.4
chevron==0.13.1
Click==7.0
colorama==0.3.9
cookiecutter==1.6.0
dateparser==0.7.1
docker==4.0.2
docutils==0.15.2
Flask==1.0.4
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jinja2-time==0.2.0
jmespath==0.9.4
jsonschema==2.6.0
MarkupSafe==1.1.1
poyo==0.5.0
pyasn1==0.4.7
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
regex==2019.8.19
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
serverlessrepo==0.1.9
six~=1.11.0
tzlocal==2.0.0
urllib3==1.25.3
websocket-client==0.56.0
Werkzeug==0.15.6
whichcraft==0.6.1
